Add support to Moodle 4.5 (adapt to MDL-82427)
