+++ version.php
@@ -25,8 +25,8 @@
 
 defined('MOODLE_INTERNAL') || die();
 
-$plugin->version   = 2024013101;          // The current plugin version (Date: YYYYMMDDXX).
+$plugin->version   = 2024100801;          // The current plugin version (Date: YYYYMMDDXX).
 $plugin->requires  = 2021051700;          // Requires this Moodle version.
 $plugin->component = 'filter_multilang2'; // Full name of the plugin (used for diagnostics).
-$plugin->release   = '2.0.1';
+$plugin->release   = '2.0.2';
 $plugin->maturity  = MATURITY_STABLE;
